moved switch location
